add constraint for MutBorrowField
